+++ Code/Framework/AzCore/AzCore/azcore_files.cmake
@@ -493,6 +493,7 @@ set(FILES
     RTTI/ReflectionManager.h
     RTTI/ReflectionManager.cpp
     RTTI/RTTI.h
+    RTTI/RTTIMacros.h
     RTTI/TemplateInfo.h
     RTTI/TypeInfo.h
     RTTI/TypeInfo.cpp
